+++ staging/src/k8s.io/kubelet/config/v1beta1/types.go
@@ -418,7 +418,7 @@ type KubeletConfiguration struct {
 	// Actual CPU utilization is much lower then pod CPU request quota.
 	// Default: 1
 	// +optional
-	MemoryAllocationRatio float32 `json:"memoryAllocationRatio ,omitempty"`
+	MemoryAllocationRatio uint64 `json:"memoryAllocationRatio ,omitempty"`
 	// qosReserved is a set of resource name to percentage pairs that specify
 	// the minimum percentage of a resource reserved for exclusive use by the
 	// guaranteed QoS tier.
